Update requests_csv.py
When there is no Header (as some of the files of CBOE) this function will crash your algo when reading a headerless csv file. Although it's a known problem in Requests it's not solved for years so we need to patch this file here. See also description in ref: https://github.com/quantopian/zipline/issues/1837)